test/CMakeLists.txt: specifically link libiio to tests
In certain conditions/linker configuration, libiio needs to be explicitly
be specified.
Otherwise underlinking won't work too well.

This seems to happen when building libad9361-iio with a different install
prefix.

And from the looks of it, in appveyor we also need to specify this for
stuff to link to tests.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>