CI/travis/before_install_linux: use `dpkg` for local install
The `apt` tool is not completely honest when installing local deb packages.

If you tell it to install a local deb package, and it finds a similar
package in the `apt` repository, it may choose to install that instead of
the local one.

It's not sure why this happens, but what's sure is sure: the `dpkg` tool
will install (or fail to install) the local libiio deb package.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>