Fix building outside git repository
Currently when building the project on a machine that has git installed,
but the source is not inside a git repository cmake fails with the
following error.

	CMake Error at CMakeLists.txt:35 (if):
	  if given arguments:

		"STREQUAL" "/home/lars/libad9361-iio"

This is due to LIBAD9361_GIT_REPO variable being empty. In order to handle
empty variables in CMake in a string comparison they need to be
encapsulated in quotes.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>